refactor: separate git commands into individual files
